Added Phone,Computer

+++ src/main/java/team17cng457/webservice/Repository/AddfeatureRepository.java
@@ -3,6 +3,6 @@ package team17cng457.webservice.Repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import team17cng457.webservice.JPA.entity.AdditionalFeature;
 
-public interface AddfeatureRepository extends JpaRepository<AdditionalFeature,Double> {
+public interface AddfeatureRepository extends JpaRepository<AdditionalFeature,Long> {
 
 }
